sysrq: add magic sysrq key to get into bootloader (fastboot) mode
If you've made it this far, you probably realize that its pretty
tough to get into a recovery mode on an ouya that doesn't boot.

Thankfully, as long as the kernel didnt panic, magic sysrq keys will
still work.

Our current best key was I which kills all running processes, if used
enough times, the system will drop itself back into recovery mode. The
problem with this is that if the system doesn't boot, say from a bad
init (which is what sysrq I kills) you won't have anything to kill and
it'll just sit there doing nothing.

This adds the magic key X which when pressed with SysRq and the Right
Alt key, will dump the device back into bootloader (fastboot) mode
for easy recovering.

It's not the magic bullet, if the kernel panics, likely this won't work
but you should be fastboot booting your kernels to test them anyway.